Resolve an Arrow data type from a string or pass through an existing type

Converts a user-friendly type name string to the corresponding Arrow DataType object. If an Arrow DataType is passed directly it is returned unchanged, so the function is safe to call on values that may already be resolved.

Usage

arrow_type(x)

Arguments

x

A string naming an Arrow type (see table above), or an Arrow DataType object.

Value

An Arrow DataType object.

Details

The following string names are recognised:

String	Arrow type

"int8"	arrow::int8()

"int16"	arrow::int16()

"int32"	arrow::int32()

"int64"	arrow::int64()

"uint8"	arrow::uint8()

"uint16"	arrow::uint16()

"uint32"	arrow::uint32()

"uint64"	arrow::uint64()

"float16"	arrow::float16()

"float32"	arrow::float32()

"float64"	arrow::float64()

"double"	arrow::float64()

"bool"	arrow::boolean()

"boolean"	arrow::boolean()

"string"	arrow::utf8()

"utf8"	arrow::utf8()

"large_string"	arrow::large_utf8()

"large_utf8"	arrow::large_utf8()

"date"	arrow::date32()

"date32"	arrow::date32()

"timestamp"	arrow::timestamp("us")

"timestamptz"	arrow::timestamp("us", timezone = "UTC")

Examples

if (FALSE) { # \dontrun{
arrow_type("int32")$ToString()    # "int32"
arrow_type("double")$ToString()   # "double"
arrow_type("date")$ToString()     # "date32"

# Passing an already-resolved type is a no-op
arrow_type(arrow::float32())$ToString()
} # }